Move NewTestGenesisWithTrieAndHeader where needed - In dot/helpers_test.go - In dot/core/helpers_test.go - In dot/digest/helpers_test.go - In dot/rpc/helpers_test.go - In dot/rpc/modules/helpers_test.go - In dot/state/helpers_test.go - In lib/grandpa/helpers_test.go - In lib/babe/helpers_test.go
