+++ dot/helpers_test.go
@@ -1,3 +1,6 @@
+// Copyright 2022 ChainSafe Systems (ON)
+// SPDX-License-Identifier: LGPL-3.0-only
+
 package dot
 
 import (
@@ -6,7 +9,11 @@ import (
 	"path/filepath"
 	"testing"
 
+	"github.com/ChainSafe/gossamer/dot/types"
+	"github.com/ChainSafe/gossamer/lib/common"
 	"github.com/ChainSafe/gossamer/lib/genesis"
+	"github.com/ChainSafe/gossamer/lib/trie"
+	"github.com/ChainSafe/gossamer/lib/utils"
 	"github.com/stretchr/testify/require"
 )
 
@@ -17,4 +24,25 @@ func writeGenesisToTestJSON(t *testing.T, genesis genesis.Genesis) (filename str
 	err = os.WriteFile(filename, jsonData, os.ModePerm)
 	require.NoError(t, err)
 	return filename
+}
+
+func newTestGenesisWithTrieAndHeader(t *testing.T) (
+	gen *genesis.Genesis, genesisTrie *trie.Trie, genesisHeader *types.Header) {
+	genesisPath := utils.GetGssmrV3SubstrateGenesisRawPathTest(t)
+	gen, err := genesis.NewGenesisFromJSONRaw(genesisPath)
+	require.NoError(t, err)
+
+	genesisTrie, err = genesis.NewTrieFromGenesis(gen)
+	require.NoError(t, err)
+
+	parentHash := common.NewHash([]byte{0})
+	stateRoot := genesisTrie.MustHash()
+	extrinsicRoot := trie.EmptyHash
+	const number = 0
+	digest := types.NewDigest()
+	genesisHeader, err = types.NewHeader(parentHash,
+		stateRoot, extrinsicRoot, number, digest)
+	require.NoError(t, err)
+
+	return gen, genesisTrie, genesisHeader
 }
