+++ dot/services_integration_test.go
@@ -17,7 +17,6 @@ import (
 	"github.com/ChainSafe/gossamer/internal/log"
 	"github.com/ChainSafe/gossamer/internal/pprof"
 	"github.com/ChainSafe/gossamer/lib/common"
-	"github.com/ChainSafe/gossamer/lib/genesis"
 	"github.com/ChainSafe/gossamer/lib/grandpa"
 	"github.com/ChainSafe/gossamer/lib/keystore"
 	"github.com/ChainSafe/gossamer/lib/runtime"
@@ -54,7 +53,7 @@ func newStateServiceWithoutMock(t *testing.T) *state.Service {
 	}
 	stateSrvc := state.NewService(stateConfig)
 	stateSrvc.UseMemDB()
-	genData, genTrie, genesisHeader := genesis.NewTestGenesisWithTrieAndHeader(t)
+	genData, genTrie, genesisHeader := newTestGenesisWithTrieAndHeader(t)
 	err := stateSrvc.Initialise(genData, genesisHeader, genTrie)
 	require.NoError(t, err)
