+++ lib/babe/helpers_test.go
@@ -1,3 +1,6 @@
+// Copyright 2022 ChainSafe Systems (ON)
+// SPDX-License-Identifier: LGPL-3.0-only
+
 package babe
 
 import (
@@ -23,6 +26,27 @@ func newDevGenesisWithTrieAndHeader(t *testing.T) (gen *genesis.Genesis, genesis
 
 	genesisHeader, err := types.NewHeader(common.NewHash([]byte{0}),
 		genesisTrie.MustHash(), trie.EmptyHash, 0, types.NewDigest())
+	require.NoError(t, err)
+
+	return gen, genesisTrie, genesisHeader
+}
+
+func newTestGenesisWithTrieAndHeader(t *testing.T) (
+	gen *genesis.Genesis, genesisTrie *trie.Trie, genesisHeader *types.Header) {
+	genesisPath := utils.GetGssmrV3SubstrateGenesisRawPathTest(t)
+	gen, err := genesis.NewGenesisFromJSONRaw(genesisPath)
+	require.NoError(t, err)
+
+	genesisTrie, err = genesis.NewTrieFromGenesis(gen)
+	require.NoError(t, err)
+
+	parentHash := common.NewHash([]byte{0})
+	stateRoot := genesisTrie.MustHash()
+	extrinsicRoot := trie.EmptyHash
+	const number = 0
+	digest := types.NewDigest()
+	genesisHeader, err = types.NewHeader(parentHash,
+		stateRoot, extrinsicRoot, number, digest)
 	require.NoError(t, err)
 
 	return gen, genesisTrie, genesisHeader
